requirements.txt
flask
requests
mysql-connector-python==8.0.31
